+++ .sequelizerc
