+++ config/config.cjs
@@ -1,0 +1,10 @@
+module.exports = {
+  development: {
+    username: process.env.USERNAME,
+    password: process.env.PASSWORD,
+    database: process.env.DATABASE,
+    host: process.env.HOST,
+    dialect: process.env.DIALECT,
+  },
+};
+console.log(process.env.PASSWORD);
