--- config/config.json
@@ -1,9 +1,0 @@
-{
-  "development": {
-    "username": "postgres",
-    "password": "silas2001",
-    "database": "E-library",
-    "host": "localhost",
-    "dialect": "postgres"
-  }
-}
